pyproject.toml

[tool.poetry]
name = "fastapi_test"
version = "0.1.0"
description = ""
authors = ["JBP"]

[tool.poetry.dependencies]
python = "^3.9"
gino = {extras = ["pg", "startlette"], version = "^1.0.1"}
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
alembic = "^1.7.1"
psycopg2-binary = "^2.9.1"
gino-starlette = "^0.1.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"

[tool.poetry.plugins."fastapi_test.modules"]
"users" = "fastapi_test.views.users"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
